updated project service

+++ src/main/java/pl/betse/beontime/projectservice/service/RoleService.java
@@ -59,9 +59,9 @@ public class RoleService {
         ProjectRoleEntity projectRoleEntity = roleRepository.
                 findByRoleGuid(guid).
                 orElseThrow(RoleNotFoundException::new);
-//        if (!projectRoleEntity.getProjectRateEntities().isEmpty()) {
-//            throw new RoleAssignedToRateException();
-//        }
+        if (!projectRoleEntity.getProjectRateEntities().isEmpty()) {
+            throw new RoleAssignedToRateException();
+        }
         roleRepository.delete(projectRoleEntity);
     }
